#26: Refactored communication service to singleton
