+++ App/app/src/main/java/com/github/travelplannerapp/signin/SignInContract.kt
@@ -1,14 +1,12 @@
 package com.github.travelplannerapp.signin
 
 
-import retrofit2.http.Body
-import io.reactivex.Observable
-import retrofit2.http.POST
+import com.github.travelplannerapp.communication.ServerApi
 
 
 interface SignInContract {
     interface View {
-        fun authorize(requestInterface: ServerAPI, jsonLoginRequest: String, handleResponse: (jsonString: String) -> Unit)
+        fun authorize(requestInterface: ServerApi, jsonLoginRequest: String, handleResponse: (jsonString: String) -> Unit)
         fun showSignUp()
         fun signIn(auth: String, email: String)
         fun showSnackbar(message: String)
@@ -19,10 +17,5 @@ interface SignInContract {
         fun signIn(email: String, password: String)
         fun signUp()
         fun handleLoginResponse(jsonString: String)
-    }
-
-    interface ServerAPI {
-        @POST("/authenticate")
-        fun authenticate(@Body jsonString: String): Observable<String>
     }
 }
