+++ App/app/src/main/java/com/github/travelplannerapp/signin/SignInPresenter.kt
@@ -9,14 +9,9 @@ import com.github.travelplannerapp.jsondatamodels.JsonLoginRequest
 import com.github.travelplannerapp.jsondatamodels.LOGIN_ANSWER
 import com.github.travelplannerapp.utils.PasswordUtils
 import com.google.gson.Gson
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.converter.scalars.ScalarsConverterFactory
 
 class SignInPresenter(view: SignInContract.View) : BasePresenter<SignInContract.View>(view), SignInContract.Presenter {
 
-    private var server = CommunicationService
     lateinit var email: String
 
     override fun signUp() {
@@ -24,12 +19,7 @@ class SignInPresenter(view: SignInContract.View) : BasePresenter<SignInContract.
     }
 
     override fun signIn(email: String, password: String) {
-        val requestInterface = Retrofit.Builder()
-                .baseUrl(server.getUrl())
-                .addConverterFactory(ScalarsConverterFactory.create())
-                .addConverterFactory(GsonConverterFactory.create())
-                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                .build().create(SignInContract.ServerAPI::class.java)
+        val requestInterface = CommunicationService.serverApi
 
         val pwd = PasswordUtils().hashPassword(password)
         if (pwd == null) {
