+++ App/app/src/main/java/com/github/travelplannerapp/travels/TravelsPresenter.kt
@@ -2,22 +2,14 @@ package com.github.travelplannerapp.travels
 
 import com.github.travelplannerapp.BasePresenter
 import com.github.travelplannerapp.communication.CommunicationService
-import retrofit2.Retrofit
-import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
-import retrofit2.converter.gson.GsonConverterFactory
 
 class TravelsPresenter(view: TravelsContract.View) : BasePresenter<TravelsContract.View>(view), TravelsContract.Presenter {
 
 
     private var travels = listOf<String>()
-    private var server = CommunicationService
 
     override fun loadTravels() {
-        val requestInterface = Retrofit.Builder()
-                .baseUrl(server.getUrl())
-                .addConverterFactory(GsonConverterFactory.create())
-                .addCallAdapterFactory(RxJava2CallAdapterFactory.create())
-                .build().create(TravelsContract.ServerAPI::class.java)
+        val requestInterface = CommunicationService.serverApi
 
         view.loadTravels(requestInterface, this::handleResponse)
     }
